Added integration test for removing molecule + group tests better
See https://github.com/i-VRESSE/workflow-builder/pull/62#issuecomment-1133022703
